[refactor] 1. port out 디렉토리 추가 (member, webtoon)
