+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/persistence/rdbms/AuthMemberPersistenceMapper.java
@@ -1,7 +1,7 @@
 package com.samsamohoh.webtoonsearch.adapter.persistence.rdbms;
 
 import com.samsamohoh.webtoonsearch.adapter.persistence.rdbms.entity.AuthMemberEntity;
-import com.samsamohoh.webtoonsearch.application.port.out.dto.AuthMemberResponse;
+import com.samsamohoh.webtoonsearch.application.port.out.member.dto.AuthMemberResponse;
 import org.springframework.stereotype.Component;
 
 @Component
