+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/application/port/in/webtoon/SearchWebtoonUseCase.java
@@ -1,8 +1,8 @@
 package com.samsamohoh.webtoonsearch.application.port.in.webtoon;
 
-import com.samsamohoh.webtoonsearch.application.port.in.webtoon.dto.SearchWebtoonCommand;
+import com.samsamohoh.webtoonsearch.application.port.in.webtoon.dto.SearchWebtoonRequest;
 import com.samsamohoh.webtoonsearch.application.port.in.webtoon.dto.WebtoonResult;
 
 public interface SearchWebtoonUseCase {
-    WebtoonResult searchWebtoons(SearchWebtoonCommand command);
+    WebtoonResult searchWebtoons(SearchWebtoonRequest request);
 }
