+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/application/port/in/webtoon/dto/SearchWebtoonRequest.java
@@ -3,6 +3,6 @@ package com.samsamohoh.webtoonsearch.application.port.in.webtoon.dto;
 import lombok.Value;
 
 @Value
-public class SearchWebtoonCommand {
+public class SearchWebtoonRequest {
     String query;
 }
