+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/application/service/AuthenticationService.java
@@ -3,8 +3,8 @@ package com.samsamohoh.webtoonsearch.application.service;
 import com.samsamohoh.webtoonsearch.application.domain.AuthMemberInfo;
 import com.samsamohoh.webtoonsearch.application.port.in.member.AuthenticationUseCase;
 import com.samsamohoh.webtoonsearch.application.port.in.member.dto.*;
-import com.samsamohoh.webtoonsearch.application.port.out.AuthMemberPort;
-import com.samsamohoh.webtoonsearch.application.port.out.dto.AuthMemberResponse;
+import com.samsamohoh.webtoonsearch.application.port.out.member.AuthMemberPort;
+import com.samsamohoh.webtoonsearch.application.port.out.member.dto.AuthMemberResponse;
 import com.samsamohoh.webtoonsearch.exception.AuthenticationFailedException;
 import com.samsamohoh.webtoonsearch.exception.MemberNotFoundException;
 import com.samsamohoh.webtoonsearch.exception.RegistrationFailedException;
