+++ webtoon-search-back/webtoon-search/src/main/resources/application.yaml
@@ -1,0 +1,91 @@
+# 서버 포트 설정
+server:
+  port: 8080
+
+spring:
+  application:
+    name: webtoon-search
+  datasource:
+    url:
+    driverClassName: org.postgresql.Driver
+    username: sa
+    password: password
+  jpa:
+    database-platform: org.hibernate.dialect.H2Dialect
+    hibernate:
+      ddl-auto: update
+    show-sql: false
+  h2:
+    console:
+      enabled: true
+
+aws:
+  open-search:
+    id: "admin"
+    pwd: "1234qwerA!"
+    region: "ap-northeast-2"
+    endpoint: "vpc-webtoon-search-dev-op-lljrssshk5kuqvybooeawwch6u.ap-northeast-2.es.amazonaws.com"
+    port: 443
+    protocol: "https"
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: health, info, metrics, prometheus
+  endpoint:
+    prometheus:
+      enabled: true
+    health:
+      probes:
+        enabled: true  # Kubernetes 프로브 활성화
+  health:
+    livenessstate:
+      enabled: true  # Liveness 프로브 활성화
+    readinessstate:
+      enabled: true  # Readiness 프로브 활성화
+  metrics:
+    tags:
+      application: ${spring.application.name}  # Metrics에 애플리케이션 이름 태그 추가
+  tracing:
+    sampling:
+      probability: 1.0
+  otlp:
+    tracing:
+      endpoint: # Tempo 위치
+
+#  tracing:
+#    sampling:
+#      probability: 1.0  # 트레이싱 샘플링 비율 설정 (1.0은 100% 샘플링)
+#    enabled: false
+#  zipkin:
+#    endpoint: http://opentelemetry-collector-service:9411/api/v2/spans  # Zipkin 엔드포인트 설정
+
+logging:
+  config: classpath:log4j2-spring.xml
+  pattern:
+    level: "%5p [${spring.application.name:},%X{traceId:-},%X{spanId:-}]"
+
+security:
+  oauth2:
+    client:
+      registration:
+        naver:
+          client-id: Oi2P9dITH68ua3Pz2MSe
+          client-secret: oTIDNlOaJg
+          redirect-uri: "{baseUrl}/login/oauth2/code/naver"
+          authorization-grant-type: authorization_code
+          scope:
+            - id        # 사용자 식별자 (네이버 ID가 아닌 고유 식별자)
+            - email      # 이메일
+            - nickname   # 닉네임
+            - gender     # 성별
+            - age        # 나이
+            - name       # 이름
+          client-name: Naver
+      provider:
+        naver:
+          authorization-uri: https://nid.naver.com/oauth2.0/authorize
+          token-uri: https://nid.naver.com/oauth2.0/token
+          user-info-uri: https://openapi.naver.com/v1/nid/me
+          user-name-attribute: response
